# apps/api/src/middleware/validator.ts
/**
 * Custom Validation Middleware
 * Provides standardized error responses for Zod validation errors.
 * Includes optional prompt injection detection for AI-related endpoints.
 */

import { z } from "zod";
import type { MiddlewareHandler } from "hono";
import { HTTP_HEADERS, CONTEXT_KEYS } from "@blueprint/shared";
import { ErrorType, createErrorJson } from "../errors";
import {
  API_HEADERS,
  HTTP_STATUS,
  VALIDATION_MESSAGES,
  ERROR_MESSAGES,
  ERROR_CODES,
  INJECTION_ERROR_MESSAGE,
  CONTENT_TYPE_NONE,
} from "../config/constants";
import { detectInjectionPatterns } from "../config/prompt-security";
import type { AppVariables } from "../types";

/**
 * Custom Zod validator that returns standardized error responses
 * Includes content-type validation for JSON requests
 */
export const validateJson = <T extends z.ZodTypeAny>(
  schema: T
): MiddlewareHandler<{
  Variables: AppVariables & {
    validatedData: z.infer<T>;
  };
}> => {
  return async (c, next) => {
    const requestId = c.get(CONTEXT_KEYS.REQUEST_ID);
    const contentType = c.req.header(API_HEADERS.REQUEST.CONTENT_TYPE);
    if (!contentType?.includes(HTTP_HEADERS.CONTENT_TYPE_JSON)) {
      return c.json(
        createErrorJson(
          ErrorType.VALIDATION,
          VALIDATION_MESSAGES.CONTENT_TYPE_MISMATCH(HTTP_HEADERS.CONTENT_TYPE_JSON),
          {
            code: ERROR_CODES.VALIDATION_ERROR,
            details: {
              expected: HTTP_HEADERS.CONTENT_TYPE_JSON,
              received: contentType || CONTENT_TYPE_NONE,
            },
            requestId,
          }
        ),
        HTTP_STATUS.BAD_REQUEST
      );
    }

    try {
      const body = await c.req.json();
      const result = schema.safeParse(body);

      if (!result.success) {
        const errorResponse = createErrorJson(ErrorType.VALIDATION, ERROR_MESSAGES.VALIDATION, {
          code: ERROR_CODES.VALIDATION_ERROR,
          details: {
            issues: result.error.issues.map((issue) => ({
              path: issue.path,
              message: issue.message,
            })),
          },
          requestId,
        });

        return c.json(errorResponse, HTTP_STATUS.BAD_REQUEST);
      }

      // Attach validated data to the context
      c.set(CONTEXT_KEYS.VALIDATED_DATA, result.data);
      await next();
    } catch {
      return c.json(
        createErrorJson(ErrorType.VALIDATION, VALIDATION_MESSAGES.INVALID_JSON_BODY, {
          code: ERROR_CODES.VALIDATION_ERROR,
          requestId,
        }),
        HTTP_STATUS.BAD_REQUEST
      );
    }
  };
};

/**
 * Field specifications for prompt injection validation.
 * Defines which fields in a validated object should be checked for
 * prompt injection patterns. Used by validatePromptInjection middleware.
 */
export interface PromptInjectionField {
  /** Dot-separated path to the field (e.g., "description", "techStack.name") */
  path: string;
  /** Human-readable label for error messages */
  label: string;
}

/**
 * Middleware that validates input fields for prompt injection patterns.
 * Composable after validateJson — runs injection detection on specified
 * fields of the already-validated request body.
 *
 * This provides defense-in-depth at the API boundary, rejecting requests
 * that contain prompt injection attempts BEFORE they reach the AI service.
 *
 * @param fields - Array of field specifications to check for injection
 * @returns Middleware handler that rejects on injection detection
 *
 * @example
 * ```typescript
 * router.post("/generate",
 *   validateJson(BlueprintRequestSchema),
 *   validatePromptInjection([
 *     { path: "description", label: "description" },
 *     { path: "projectName", label: "project name" },
 *   ]),
 *   generateController.generateBlueprint
 * );
 * ```
 */
export const validatePromptInjection = (
  fields: readonly PromptInjectionField[]
): MiddlewareHandler => {
  return async (c, next) => {
    const requestId = c.get(CONTEXT_KEYS.REQUEST_ID) as string | undefined;
    const data = c.get(CONTEXT_KEYS.VALIDATED_DATA) as Record<string, unknown>;
    if (!data) {
      await next();
      return;
    }

    for (const field of fields) {
      const value = getNestedValue(data, field.path);
      if (typeof value === "string" && value.length > 0) {
        const detected = detectInjectionPatterns(value);
        if (detected.length > 0) {
          return c.json(
            createErrorJson(ErrorType.VALIDATION, ERROR_MESSAGES.VALIDATION, {
              code: ERROR_CODES.VALIDATION_ERROR,
              details: {
                field: field.path,
                message: INJECTION_ERROR_MESSAGE(field.label),
              },
              requestId,
            }),
            HTTP_STATUS.BAD_REQUEST
          );
        }
      }
      if (Array.isArray(value)) {
        for (const element of value) {
          const injectionResult = checkValueForInjection(element);
          if (injectionResult) {
            return c.json(
              createErrorJson(ErrorType.VALIDATION, ERROR_MESSAGES.VALIDATION, {
                code: ERROR_CODES.VALIDATION_ERROR,
                details: {
                  field: field.path,
                  message: INJECTION_ERROR_MESSAGE(field.label),
                },
                requestId,
              }),
              HTTP_STATUS.BAD_REQUEST
            );
          }
        }
      }
    }

    await next();
  };
};

function getNestedValue(obj: Record<string, unknown>, path: string): unknown {
  return path.split(".").reduce<unknown>((current, key) => {
    if (current && typeof current === "object" && key in current) {
      return (current as Record<string, unknown>)[key];
    }
    return undefined;
  }, obj);
}

/**
 * Checks a value for prompt injection patterns.
 * Handles strings directly, and recursively checks all string
 * values within objects — enabling array-of-objects validation.
 *
 * @param value - The value to check (string, object, or primitive)
 * @returns The first detected pattern label, or null if clean
 */
function checkValueForInjection(value: unknown): string | null {
  if (typeof value === "string" && value.length > 0) {
    const detected = detectInjectionPatterns(value);
    return detected.length > 0 ? (detected[0] ?? null) : null;
  }
  if (value && typeof value === "object" && !Array.isArray(value)) {
    for (const val of Object.values(value as Record<string, unknown>)) {
      const result = checkValueForInjection(val);
      if (result) return result;
    }
  }
  return null;
}