# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Blueprint Generator API
  version: 1.0.0
  description: |
    Cloudflare Workers backend for AI-powered project architecture documentation generation.
    Provides streaming SSE endpoints for generating blueprints and tasks, content refinement,
    data export/import, storage tracking, and blueprint sharing.
  contact:
    name: Blueprintify Team
    url: https://github.com/cpa03/blueprintify
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://blueprintify.cpa03-cmz.workers.dev
    description: Production Cloudflare Workers Edge
  - url: http://localhost:8787
    description: Local API Development Server
  - url: http://localhost:3000/api
    description: Local Web Proxy Server
security:
  - ApiKeyAuth: []
paths:
...
  /generate:
    post:
      summary: Generate Blueprint
      description: Generates comprehensive project architecture documentation (`blueprint.md`) from project configuration using OpenAI streaming (SSE).
      operationId: generateBlueprint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlueprintRequest"
      responses:
        "200":
          description: Server-Sent Events stream containing generated markdown chunks
          headers:
            Content-Type:
              schema:
                type: string
                example: text/event-stream
            Cache-Control:
              schema:
                type: string
                example: no-cache
            Connection:
              schema:
                type: string
                example: keep-alive
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/StreamChunk"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "429":
          $ref: "#/components/responses/RateLimitError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailableError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API Key authentication header. Default role is 'user'; key matching ADMIN_API_KEY provides 'admin' role.
  responses:
    ValidationError:
      description: Request validation failed (400 Bad Request)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthenticationError:
      description: Missing or invalid API key credentials (401 Unauthorized)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimitError:
      description: Rate limit exceeded (429 Too Many Requests)
      headers:
        Retry-After:
          schema:
            type: integer
            description: Seconds until rate limit window resets
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error (500 Internal Server Error)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailableError:
      description: Service temporarily unavailable or configuration error (503 Service Unavailable)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required:
            - type
            - message
            - timestamp
          properties:
            type:
              type: string
              enum:
                - validation
                - authentication
                - authorization
                - not_found
                - configuration
                - network
                - ai_service
                - service_unavailable
                - internal
            message:
              type: string
            code:
              type: string
            details:
              type: object
              description: >-
                Error-specific context. For 400 validation failures this carries an
                `issues` array of ValidationIssue objects; prompt-injection rejections
                carry `field` and `message`.
              properties:
                issues:
                  type: array
                  items:
                    $ref: "#/components/schemas/ValidationIssue"
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              description: >-
                Correlation id for this request, identical to the value logged for the
                same request and to the `X-Request-ID` response header. Format is
                `{epochMilliseconds}-{random}`, NOT a UUID. It is not present on every
                error body: the global error handler (404, 500, 504 and the
                circuit-breaker 503), API-key authentication (401, 503), request-body
                validation (400) and the `/share` and `/storage` routes emit it, while
                authorization (401, 403), rate limiting (429 and its 503 configuration
                error), the payload-size 413 and the `/import` route's own validation
                and internal errors do not.
              pattern: '^\d+-[a-z0-9]+$'
              example: "1790884637380-om7wd1rqu"
    ValidationIssue:
      type: object
      required:
        - path
        - message
      properties:
        path:
          type: array
          description: >-
            Path to the offending value as an ordered array of segments, NOT a dotted
            string. Object keys appear as strings and array indices as integers, so a
            bad `techStack[2].category` value yields ["techStack", 2, "category"].
          items:
            oneOf:
              - type: string
              - type: integer
          example: ["techStack", 2, "category"]
        message:
          type: string
          example: "Expected 'frontend' | 'backend' | ..."
    TechStackItem:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        category:
          type: string
          enum:
            - frontend
            - backend
            - database
            - hosting
            - ai
            - testing
            - styling
            - other
        subcategory:
          type: string
          enum:
            - relational
            - nosql
            - vector
            - graph
            - edge
            - search
            - cache
            - serverless
        version:
          type: string
          maxLength: 50
        description:
          type: string
          maxLength: 2000
        features:
          type: array
          items:
            type: string
            maxLength: 200
    BlueprintRequest:
      type: object
      required:
        - projectName
        - description
        - techStack
      properties:
        projectName:
          type: string
          minLength: 1
          maxLength: 100
          example: "E-Commerce Microservices"
        description:
          type: string
          minLength: 10
          maxLength: 2000
          example: "A high-performance modular e-commerce architecture with real-time cart synchronization."
        techStack:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/TechStackItem"
        features:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 200
        targetAudience:
          type: string
          maxLength: 500
        constraints:
          type: string
          maxLength: 1000
...
    StreamChunk:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [content, error, done]
        content:
          type: string
        error:
          type: string
...